change data type of show id to small int size

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS shows (
-    id BIGSERIAL PRIMARY KEY,
+    id SMALLSERIAL PRIMARY KEY,
     title TEXT NOT NULL,
     desired_format TEXT,
     desired_folder TEXT,
@@ -19,8 +19,8 @@ CREATE TABLE IF NOT EXISTS shows (
 );
 
 CREATE TABLE IF NOT EXISTS show_entry (
-    id BIGSERIAL PRIMARY KEY,
-    show_id BIGSERIAL NOT NULL REFERENCES shows(id),
+    id SMALLSERIAL PRIMARY KEY,
+    show_id SMALLINT NOT NULL REFERENCES shows(id),
     episode SMALLINT NOT NULL,
     current_state show_state NOT NULL DEFAULT 'downloading',
     torrent_hash TEXT NOT NULL,
